fix: order month widget items by occurrence
