+++ app/src/main/java/com/kgs/calendar/widget/KgsWidgetProvider.kt
@@ -27,7 +27,6 @@ import android.text.SpannableString
 import android.text.Spanned
 import android.text.style.StyleSpan
 import android.util.Log
-import android.util.SizeF
 import android.util.TypedValue
 import android.view.Gravity
 import android.view.View
@@ -87,19 +86,7 @@ private const val EXTRA_COLLECTION_ACTION = "kgs_widget_collection_action"
 private const val COLLECTION_ACTION_OPEN = "open"
 private const val COLLECTION_ACTION_TOGGLE_TASK = "toggle_task"
 private const val COLLECTION_ACTION_TOGGLE_SUBTASKS = "toggle_subtasks"
-private const val WIDGET_MONTH_MAX_LANES = 10
-private const val WIDGET_MONTH_MAX_TEXT_LANES = 10
-private const val WIDGET_MONTH_VERTICAL_CHROME_DP = 67
-private const val WIDGET_MONTH_TEXT_FIXED_CHROME_DP = 15
-private const val WIDGET_MONTH_TEXT_ROW_STRIDE_DP = 15
-private const val WIDGET_MONTH_MIN_TEXT_CELL_HEIGHT_DP = 46
-private const val WIDGET_MONTH_MINI_FIXED_HEIGHT_DP = 19
-private const val WIDGET_MONTH_MINI_DOT_ROW_HEIGHT_DP = 9
-private const val WIDGET_MONTH_DOTS_PER_ROW = 5
-private const val WIDGET_MONTH_MAX_DOTS = 10
-private const val WIDGET_MONTH_DOT_SLOT_WIDTH_DP = 10f
 private const val WIDGET_MONTH_DOT_SIZE_DP = 3.5f
-private const val WIDGET_MONTH_CELL_HORIZONTAL_MARGIN_DP = 2f
 private const val WIDGET_MONTH_SPAN_CHIP_HEIGHT_DP = 14f
 private const val WIDGET_MONTH_SPAN_CHIP_HORIZONTAL_INSET_DP = 2f
 private const val WIDGET_MONTH_SPAN_FADE_WIDTH_DP = 14f
@@ -5444,421 +5431,6 @@ internal class KgsWidgetDataSource(
 
     private fun LocalTime.timeText(): String =
         format(DateTimeFormatter.ofPattern("HH:mm"))
-}
-
-internal object WidgetMonthModel {
-    fun gridStart(month: YearMonth, firstDayOfWeek: DayOfWeek): LocalDate {
-        val first = month.atDay(1)
-        val offset = (first.dayOfWeek.value - firstDayOfWeek.value + 7) % 7
-        return first.minusDays(offset.toLong())
-    }
-
-    fun rowCount(month: YearMonth, firstDayOfWeek: DayOfWeek): Int {
-        val leadingDays = (month.atDay(1).dayOfWeek.value - firstDayOfWeek.value + 7) % 7
-        return maxOf(5, (leadingDays + month.lengthOfMonth() + 6) / 7)
-    }
-
-    fun page(
-        month: YearMonth,
-        start: LocalDate,
-        rowCount: Int,
-        monthLayout: WidgetMonthLayout,
-    ): WidgetMonthPage {
-        val cells = (0 until 42).map { offset ->
-            val date = start.plusDays(offset.toLong())
-            WidgetMonthCellContent(
-                date = date,
-                inCurrentMonth = YearMonth.from(date) == month,
-                items = monthLayout.itemsByDay[date].orEmpty(),
-                totalItemCount = monthLayout.totalByDay[date] ?: 0,
-            )
-        }
-        return WidgetMonthPage(month, rowCount, cells)
-    }
-
-    fun layout(
-        month: YearMonth,
-        gridStart: LocalDate,
-        rowCount: Int,
-        candidates: List<WidgetMonthCandidate>,
-        locale: Locale,
-    ): WidgetMonthLayout {
-        val itemsByDay = mutableMapOf<LocalDate, MutableList<WidgetMonthItem>>()
-        val totalByDay = mutableMapOf<LocalDate, Int>()
-
-        repeat(rowCount) { row ->
-            val rowStartDate = gridStart.plusDays((row * 7).toLong())
-            val rowEndDate = rowStartDate.plusDays(6)
-            val rowDays = (0 until 7)
-                .map { rowStartDate.plusDays(it.toLong()) }
-                .filter { YearMonth.from(it) == month }
-            if (rowDays.isEmpty()) return@repeat
-            val rowStart = rowDays.first()
-            val rowEnd = rowDays.last()
-            val rowCandidates = candidates
-                .filter { it.end >= rowStart && it.start <= rowEnd }
-                .distinctBy { it.id }
-            rowDays.forEach { day ->
-                totalByDay[day] = rowCandidates.count { day in it.start..it.end }
-            }
-
-            data class MutableSegment(
-                val candidate: WidgetMonthCandidate,
-                val visualStart: LocalDate,
-                val visualEnd: LocalDate,
-                val lane: Int,
-            )
-
-            val laneSegments = mutableListOf<MutableList<MutableSegment>>()
-            val segments = mutableListOf<MutableSegment>()
-            val sortedCandidates = rowCandidates.sortedWith(
-                compareBy<WidgetMonthCandidate> { maxOf(it.start, rowStart) }
-                    .thenByDescending { it.spanDays }
-                    .thenBy { it.title.lowercase(locale) },
-            )
-
-            sortedCandidates.forEach { candidate ->
-                fun doesNotOverlap(existing: MutableSegment): Boolean =
-                    existing.candidate.end < candidate.start || candidate.end < existing.candidate.start
-
-                val existingLane = laneSegments.indexOfFirst { lane -> lane.all(::doesNotOverlap) }
-                val targetLane = if (existingLane >= 0) existingLane else laneSegments.size
-                val segment = MutableSegment(
-                    candidate = candidate,
-                    visualStart = maxOf(candidate.start, rowStart),
-                    visualEnd = minOf(candidate.end, rowEnd),
-                    lane = targetLane,
-                )
-                if (targetLane < laneSegments.size) {
-                    laneSegments[targetLane] += segment
-                } else {
-                    laneSegments += mutableListOf(segment)
-                }
-                segments += segment
-            }
-
-            fun WidgetMonthCandidate.visuallyConnectsTo(other: WidgetMonthCandidate): Boolean =
-                color == other.color && title.trim().equals(other.title.trim(), ignoreCase = true)
-
-            segments
-                .filter { it.lane < WIDGET_MONTH_MAX_LANES && !it.visualEnd.isBefore(it.visualStart) }
-                .forEach { segment ->
-                    var day = segment.visualStart
-                    while (!day.isAfter(segment.visualEnd)) {
-                        val previousSameLaneSegment = segments.firstOrNull { other ->
-                            other !== segment &&
-                                other.lane == segment.lane &&
-                                other.visualEnd == day.minusDays(1) &&
-                                other.candidate.visuallyConnectsTo(segment.candidate)
-                        }
-                        val nextSameLaneSegment = segments.firstOrNull { other ->
-                            other !== segment &&
-                                other.lane == segment.lane &&
-                                other.visualStart == day.plusDays(1) &&
-                                segment.candidate.visuallyConnectsTo(other.candidate)
-                        }
-                        val continuesFromPrevious = segment.candidate.start < day || previousSameLaneSegment != null
-                        val continuesToNext = segment.candidate.end > day || nextSameLaneSegment != null
-                        val fadesFromPrevious = day == segment.visualStart && segment.candidate.start < segment.visualStart
-                        val fadesToNext = day == segment.visualEnd && segment.candidate.end > segment.visualEnd
-                        itemsByDay.getOrPut(day) { mutableListOf() } += WidgetMonthItem(
-                            id = segment.candidate.id,
-                            title = if (day == segment.visualStart && previousSameLaneSegment == null) segment.candidate.title else "",
-                            color = segment.candidate.color,
-                            sortMillis = segment.candidate.sortMillis,
-                            lane = segment.lane,
-                            continuesFromPrevious = continuesFromPrevious,
-                            continuesToNext = continuesToNext,
-                            fadesFromPrevious = fadesFromPrevious,
-                            fadesToNext = fadesToNext,
-                            completed = segment.candidate.completed,
-                        )
-                        day = day.plusDays(1)
-                    }
-                }
-        }
-
-        return WidgetMonthLayout(
-            itemsByDay = itemsByDay.mapValues { (_, items) ->
-                items.sortedWith(compareBy<WidgetMonthItem> { it.lane }.thenBy { it.sortMillis }.thenBy { it.title.lowercase(locale) })
-            },
-            totalByDay = totalByDay,
-        )
-    }
-}
-
-internal enum class WidgetSizeBucket(
-    val remoteSize: SizeF,
-    val layoutRes: Int,
-    val textItemCapacity: Int,
-    val miniDotCapacity: Int,
-    val rootPaddingDp: Int,
-    val titleTextSp: Float,
-    val weekdayTextSp: Float,
-    val dayTextSp: Float,
-    val chipTextSp: Float,
-) {
-    Tiny(SizeF(170f, 220f), R.layout.widget_month_cell_mini, 0, 3, 5, 13f, 7.5f, 9f, 0f),
-    Mini(SizeF(220f, 285f), R.layout.widget_month_cell_mini, 0, 6, 6, 14f, 8f, 10f, 0f),
-    Compact(SizeF(260f, 350f), R.layout.widget_month_cell_dynamic, WIDGET_MONTH_MAX_TEXT_LANES, 0, 6, 15f, 8.5f, 10f, 9f),
-    Standard(SizeF(300f, 390f), R.layout.widget_month_cell_dynamic, WIDGET_MONTH_MAX_TEXT_LANES, 0, 6, 15f, 8.5f, 10f, 9f),
-    Comfortable(SizeF(320f, 470f), R.layout.widget_month_cell_dynamic, WIDGET_MONTH_MAX_TEXT_LANES, 0, 6, 15f, 8.5f, 10f, 9f),
-    Expanded(SizeF(360f, 560f), R.layout.widget_month_cell_dynamic, WIDGET_MONTH_MAX_TEXT_LANES, 0, 6, 15f, 8.5f, 10f, 9f),
-    Max(SizeF(360f, 680f), R.layout.widget_month_cell_dynamic, WIDGET_MONTH_MAX_TEXT_LANES, 0, 6, 15f, 8.5f, 10f, 9f);
-
-    val usesDotCells: Boolean
-        get() = textItemCapacity == 0
-
-    fun visibleItemCount(totalCount: Int): Int =
-        minOf(totalCount.coerceAtLeast(0), if (usesDotCells) miniDotCapacity else textItemCapacity)
-
-    fun overflowCount(totalCount: Int): Int =
-        (totalCount.coerceAtLeast(0) - visibleItemCount(totalCount)).coerceAtLeast(0)
-
-    companion object {
-        fun from(size: WidgetSize): WidgetSizeBucket = from(size, rowCount = 5)
-
-        fun from(size: WidgetSize, rowCount: Int): WidgetSizeBucket {
-            val rows = rowCount.coerceIn(5, 6)
-            val availableHeight = (size.heightDp - WIDGET_MONTH_VERTICAL_CHROME_DP).coerceAtLeast(0)
-            val weekCellHeight = availableHeight / rows
-            val dayWidth = size.widthDp / 7
-            return when {
-                dayWidth < 30 || weekCellHeight < 29 -> Tiny
-                dayWidth < 35 || weekCellHeight < 42 -> Mini
-                weekCellHeight < 58 -> Compact
-                weekCellHeight < 74 -> Standard
-                weekCellHeight < 92 -> Comfortable
-                weekCellHeight < 118 -> Expanded
-                else -> Max
-            }
-        }
-    }
-}
-
-internal data class WidgetMonthRenderSpec(
-    val bucket: WidgetSizeBucket,
-    val weekCellHeightDp: Int,
-    val cellContentWidthDp: Float = WIDGET_MONTH_DOT_SLOT_WIDTH_DP * WIDGET_MONTH_DOTS_PER_ROW,
-    val weekContentWidthDp: Float = cellContentWidthDp * 7f,
-) {
-    val layoutRes: Int get() = if (usesDotCells) R.layout.widget_month_cell_mini else bucket.layoutRes
-    val baseTextItemCapacity: Int get() = bucket.textItemCapacity
-    val rootPaddingDp: Int get() = bucket.rootPaddingDp
-    val titleTextSp: Float get() = bucket.titleTextSp
-    val weekdayTextSp: Float get() = bucket.weekdayTextSp
-    val dayTextSp: Float get() = bucket.dayTextSp
-    val chipTextSp: Float get() = bucket.chipTextSp
-    val totalWidthDp: Float get() = weekContentWidthDp + rootPaddingDp * 2f
-    val usesDotCells: Boolean get() = bucket.usesDotCells || weekCellHeightDp < WIDGET_MONTH_MIN_TEXT_CELL_HEIGHT_DP || textRowsThatFit() <= 0
-
-    fun textItemCapacityFor(rowCells: List<WidgetMonthCellContent>): Int {
-        if (usesDotCells) return 0
-        val candidateCapacity = minOf(baseTextItemCapacity, textRowsThatFit())
-        if (candidateCapacity <= 0) return 0
-        return candidateCapacity
-    }
-
-    fun miniDotCapacityFor(totalItemCount: Int): Int {
-        if (!usesDotCells || totalItemCount <= 0) return 0
-        val rowsThatFit = miniDotRowsThatFit()
-        return minOf(WIDGET_MONTH_MAX_DOTS, rowsThatFit * miniDotsPerRow(), totalItemCount)
-    }
-
-    fun miniDotRowCountsFor(totalItemCount: Int): Pair<Int, Int> {
-        val visibleCount = miniDotCapacityFor(totalItemCount)
-        if (visibleCount <= 0) return 0 to 0
-        val dotsPerRow = miniDotsPerRow()
-        if (dotsPerRow <= 0 || miniDotRowsThatFit() <= 0) return 0 to 0
-        if (miniDotRowsThatFit() == 1 || visibleCount <= dotsPerRow) {
-            return visibleCount to 0
-        }
-        val secondRow = (visibleCount / 2).coerceAtMost(dotsPerRow)
-        val firstRow = (visibleCount - secondRow).coerceAtMost(dotsPerRow)
-        return firstRow to secondRow
-    }
-
-    fun miniDotsPerRow(): Int {
-        val safeWidth = cellContentWidthDp.coerceAtLeast(0f)
-        return (safeWidth / WIDGET_MONTH_DOT_SLOT_WIDTH_DP).toInt().coerceIn(0, WIDGET_MONTH_DOTS_PER_ROW)
-    }
-
-    fun spanChipWidthDp(span: Int): Float =
-        (weekContentWidthDp / 7f * span.coerceIn(1, 7)).coerceAtLeast(1f)
-
-    fun showMiniOverflow(hiddenItemCount: Int): Boolean =
-        hiddenItemCount > 0
-
-    fun showTextOverflow(textItemCapacity: Int): Boolean =
-        textItemCapacity > 0
-
-    private fun textRowsThatFit(): Int =
-        ((weekCellHeightDp - WIDGET_MONTH_TEXT_FIXED_CHROME_DP) / WIDGET_MONTH_TEXT_ROW_STRIDE_DP)
-            .coerceIn(0, WIDGET_MONTH_MAX_TEXT_LANES + 1)
-
-    private fun miniDotRowsThatFit(): Int =
-        ((weekCellHeightDp - WIDGET_MONTH_MINI_FIXED_HEIGHT_DP) / WIDGET_MONTH_MINI_DOT_ROW_HEIGHT_DP)
-            .coerceIn(0, 2)
-
-    companion object {
-        fun from(size: WidgetSize, rowCount: Int): WidgetMonthRenderSpec {
-            val bucket = WidgetSizeBucket.from(size, rowCount)
-            val weekContentWidth = weekContentWidthDp(size, bucket)
-            return WidgetMonthRenderSpec(
-                bucket = bucket,
-                weekCellHeightDp = weekCellHeightDp(size, rowCount),
-                cellContentWidthDp = (weekContentWidth / 7f - WIDGET_MONTH_CELL_HORIZONTAL_MARGIN_DP).coerceAtLeast(0f),
-                weekContentWidthDp = weekContentWidth,
-            )
-        }
-
-        fun weekCellHeightDp(size: WidgetSize, rowCount: Int): Int {
-            val rows = rowCount.coerceIn(5, 6)
-            return (size.heightDp - WIDGET_MONTH_VERTICAL_CHROME_DP).coerceAtLeast(0) / rows
-        }
-
-        fun cellContentWidthDp(size: WidgetSize, bucket: WidgetSizeBucket): Float {
-            val gridWidth = weekContentWidthDp(size, bucket)
-            return (gridWidth / 7f - WIDGET_MONTH_CELL_HORIZONTAL_MARGIN_DP).coerceAtLeast(0f)
-        }
-
-        fun weekContentWidthDp(size: WidgetSize, bucket: WidgetSizeBucket): Float {
-            val gridWidth = (size.widthDp - bucket.rootPaddingDp * 2).coerceAtLeast(0)
-            return gridWidth.toFloat().coerceAtLeast(0f)
-        }
-    }
-}
-
-internal data class WidgetMonthPage(
-    val month: YearMonth,
-    val rowCount: Int,
-    val cells: List<WidgetMonthCellContent>,
-) {
-    fun title(locale: Locale): String =
-        month.format(DateTimeFormatter.ofPattern("MMMM yyyy", locale))
-}
-
-internal data class WidgetMonthCellContent(
-    val date: LocalDate,
-    val inCurrentMonth: Boolean,
-    val items: List<WidgetMonthItem>,
-    val totalItemCount: Int,
-)
-
-internal data class WidgetMonthLayout(
-    val itemsByDay: Map<LocalDate, List<WidgetMonthItem>>,
-    val totalByDay: Map<LocalDate, Int>,
-)
-
-internal data class WidgetMonthCandidate(
-    val id: String,
-    val title: String,
-    val color: Int,
-    val sortMillis: Long,
-    val start: LocalDate,
-    val end: LocalDate,
-    val completed: Boolean,
-) {
-    val spanDays: Long = ChronoUnit.DAYS.between(start, end).coerceAtLeast(0) + 1
-}
-
-internal data class WidgetMonthItem(
-    val id: String,
-    val title: String,
-    val color: Int,
-    val sortMillis: Long,
-    val lane: Int,
-    val continuesFromPrevious: Boolean,
-    val continuesToNext: Boolean,
-    val fadesFromPrevious: Boolean,
-    val fadesToNext: Boolean,
-    val completed: Boolean,
-)
-
-internal data class WidgetMonthWeekSegment(
-    val startColumn: Int,
-    val endColumn: Int,
-    val title: String,
-    val item: WidgetMonthItem,
-) {
-    val columnSpan: Int = endColumn - startColumn + 1
-}
-
-internal fun List<WidgetMonthCellContent>.monthWeekSegments(lane: Int): List<WidgetMonthWeekSegment> {
-    val segments = mutableListOf<WidgetMonthWeekSegment>()
-    var column = 0
-    while (column < size) {
-        if (!this[column].inCurrentMonth) {
-            column += 1
-            continue
-        }
-        val firstItem = this[column].items.firstOrNull { it.lane == lane } ?: run {
-            column += 1
-            continue
-        }
-
-        var endColumn = column
-        var last = firstItem
-        while (endColumn + 1 < size) {
-            val next = this[endColumn + 1].items.firstOrNull { it.lane == lane }
-            if (next == null || !this[endColumn + 1].inCurrentMonth || !last.visuallyContinuesInto(next)) {
-                break
-            }
-            endColumn += 1
-            last = next
-        }
-
-        val title = (column..endColumn)
-            .asSequence()
-            .mapNotNull { index -> this[index].items.firstOrNull { it.lane == lane }?.title?.takeIf(String::isNotBlank) }
-            .firstOrNull()
-            .orEmpty()
-        segments += WidgetMonthWeekSegment(
-            startColumn = column,
-            endColumn = endColumn,
-            title = title,
-            item = firstItem.copy(
-                title = title,
-                continuesToNext = last.continuesToNext,
-                fadesToNext = last.fadesToNext,
-            ),
-        )
-        column = endColumn + 1
-    }
-    return segments
-}
-
-internal fun WidgetMonthItem.visuallyContinuesInto(next: WidgetMonthItem): Boolean =
-    lane == next.lane &&
-        color == next.color &&
-        continuesToNext &&
-        next.continuesFromPrevious
-
-internal fun List<WidgetMonthCellContent>.monthBottomFadeSegments(textItemCapacity: Int): List<WidgetMonthWeekSegment> {
-    if (textItemCapacity <= 0) return emptyList()
-    val bottomLane = textItemCapacity - 1
-    val candidates = monthWeekSegments(bottomLane)
-        .asSequence()
-        .filter { segment ->
-            val item = segment.item
-            segment.columnSpan > 1 ||
-                item.continuesFromPrevious ||
-                item.continuesToNext ||
-                item.fadesFromPrevious ||
-                item.fadesToNext
-        }
-        .sortedWith(compareBy<WidgetMonthWeekSegment> { it.startColumn }.thenByDescending { it.columnSpan })
-        .toList()
-
-    val result = mutableListOf<WidgetMonthWeekSegment>()
-    var nextFreeColumn = 0
-    candidates.forEach { segment ->
-        if (segment.startColumn >= nextFreeColumn) {
-            result += segment
-            nextFreeColumn = segment.endColumn + 1
-        }
-    }
-    return result
 }
 
 private fun monthSpanChipLayout(span: Int): Int = when (span.coerceIn(1, 7)) {
